# [SQL indexer] Add functions for filtering packages at database level