--- a/internal/database/functions_test.go
+++ b/internal/database/functions_test.go
@@ -25,6 +25,8 @@ func TestSemverCompareConstraint(t *testing.T) {
 		{"1.2.3", "^1.2.0", true},
 		{"1.2.3", ">=1.2.0, <2.0.0", true},
 		{"1.2.3", ">=1.3.0", false},
+		{"2.3.4", "^1.0.0 || ^2.0.0", true},
+		{"2.3.4", "^1.0.0 || ~2.3.0", true},
 		{"1.2.3-beta", ">=1.0.0", false},  // Pre-release versions are not included unless specified
 		{"1.2.3-beta", ">=1.0.0-0", true}, // Pre-release versions are included with -0
 		{"1.2.3", "", true},               // No constraint means always true
@@ -43,25 +45,49 @@ func TestSemverCompareConstraint(t *testing.T) {
 	}
 }
 
-func TestSemverCompareOperation(t *testing.T) {
+func TestSemverCompareGreaterThanEqual(t *testing.T) {
 	tests := []struct {
 		firstVersion  string
-		operation     string
 		secondVersion string
 		expected      bool
 	}{
-		{"1.2.3", ">=", "1.0.0", true},
-		{"1.2.3", "<", "1.0.0", false},
-		{"1.2.3", ">", "1.2.0", true},
-		{"1.2.3", "<=", "1.2.3", true},
-		{"1.2.3", "!=", "1.2.4", true},
-		{"1.2.3-beta", ">=", "1.0.0", false},  // Pre-release versions are not included unless specified
-		{"1.2.3-beta", ">=", "1.0.0-0", true}, // Pre-release versions are included with -0
+		{"1.2.3", "1.0.0", true},
+		{"1.0.0", "1.2.3", false},
+		{"1.2.3-beta", "1.0.0", true}, // Pre-release versions are only "excluded" when matching against constrsaints, not in direct comparisons
+		{"1.2.3-beta", "1.0.0-0", true},
+		{"1.2.3-beta", "1.2.3", false},
+		{"1.2.3", "1.2.3-beta", true},
+		{"1.2.3-0", "1.2.3-beta", false},
 	}
 
 	for _, tt := range tests {
-		t.Run(fmt.Sprintf("%s %s %s", tt.firstVersion, tt.operation, tt.secondVersion), func(t *testing.T) {
-			result, err := semverCompareOperation(nil, []driver.Value{tt.firstVersion, tt.operation, tt.secondVersion})
+		t.Run(fmt.Sprintf("%s %s", tt.firstVersion, tt.secondVersion), func(t *testing.T) {
+			result, err := semverCompareGreaterThanEqual(nil, []driver.Value{tt.firstVersion, tt.secondVersion})
+			require.NoError(t, err)
+			assert.Equal(t, tt.expected, result)
+		})
+	}
+}
+
+func TestSemverCompareLessThanEqual(t *testing.T) {
+	tests := []struct {
+		firstVersion  string
+		secondVersion string
+		expected      bool
+	}{
+		{"1.2.3", "1.0.0", false},
+		{"1.0.0", "1.2.3", true},
+		{"1.2.3-beta", "1.0.0", false}, // Pre-release versions are only "excluded" when matching against constrsaints, not in direct comparisons
+		{"1.0.0", "1.2.3-beta", true},
+		{"1.2.3", "1.2.3-beta", false},
+		{"1.2.3-beta", "1.2.3", true},
+		{"1.2.3-beta", "1.0.0-0", false},
+		{"1.2.3-0", "1.2.3-beta", true},
+	}
+
+	for _, tt := range tests {
+		t.Run(fmt.Sprintf("%s %s", tt.firstVersion, tt.secondVersion), func(t *testing.T) {
+			result, err := semverCompareLessThanEqual(nil, []driver.Value{tt.firstVersion, tt.secondVersion})
 			require.NoError(t, err)
 			assert.Equal(t, tt.expected, result)
 		})