--- a/internal/storage/sqlindexer.go
+++ b/internal/storage/sqlindexer.go
@@ -17,6 +17,7 @@ import (
 	"time"
 
 	"cloud.google.com/go/storage"
+	"github.com/Masterminds/semver/v3"
 	"github.com/prometheus/client_golang/prometheus"
 	"go.elastic.co/apm/v2"
 	"go.uber.org/zap"
@@ -319,18 +320,25 @@ func createDatabasePackage(pkg *packages.Package, cursor string) (*database.Pack
 		kibanaVersion = pkg.Conditions.Kibana.Version
 	}
 
+	formatVersionSemver, err := semver.NewVersion(pkg.FormatVersion)
+	if err != nil {
+		return nil, fmt.Errorf("invalid format version '%s' for package %s-%s: %w", pkg.FormatVersion, pkg.Name, pkg.Version, err)
+	}
+	formatVersionMajorMinor := fmt.Sprintf("%d.%d", formatVersionSemver.Major(), formatVersionSemver.Minor())
+
 	newPackage := database.Package{
-		Cursor:        cursor,
-		Name:          pkg.Name,
-		Version:       pkg.Version,
-		FormatVersion: pkg.FormatVersion,
-		Path:          fmt.Sprintf("%s-%s.zip", pkg.Name, pkg.Version),
-		Type:          pkg.Type,
-		Release:       pkg.Release,
-		KibanaVersion: kibanaVersion,
-		Prerelease:    pkg.IsPrerelease(),
-		Data:          fullContents,
-		BaseData:      baseContents,
+		Cursor:                  cursor,
+		Name:                    pkg.Name,
+		Version:                 pkg.Version,
+		FormatVersion:           pkg.FormatVersion,
+		FormatVersionMajorMinor: formatVersionMajorMinor,
+		Path:                    fmt.Sprintf("%s-%s.zip", pkg.Name, pkg.Version),
+		Type:                    pkg.Type,
+		Release:                 pkg.Release,
+		KibanaVersion:           kibanaVersion,
+		Prerelease:              pkg.IsPrerelease(),
+		Data:                    fullContents,
+		BaseData:                baseContents,
 	}
 
 	return &newPackage, nil
@@ -365,6 +373,12 @@ func (i *SQLIndexer) Get(ctx context.Context, opts *packages.GetOptions) (packag
 			}
 			if opts.Filter.KibanaVersion != nil {
 				options.Filter.KibanaVersion = opts.Filter.KibanaVersion.String()
+			}
+			if opts.Filter.SpecMin != nil {
+				options.Filter.SpecMin = opts.Filter.SpecMin.String()
+			}
+			if opts.Filter.SpecMax != nil {
+				options.Filter.SpecMax = opts.Filter.SpecMax.String()
 			}
 		}
 		if opts != nil {