--- a/internal/storage/sqlindexer_test.go
+++ b/internal/storage/sqlindexer_test.go
@@ -6,6 +6,7 @@ package storage
 
 import (
 	"context"
+	"encoding/json"
 	"fmt"
 	"path/filepath"
 	"testing"
@@ -485,4 +486,88 @@ func TestSQLGet_IndexUpdated(t *testing.T) {
 	require.Equal(t, "1password", foundPackages[0].Name)
 	require.Equal(t, "0.2.0", foundPackages[0].Version)
 	require.Equal(t, "1Password Events Reporting UPDATED", *foundPackages[0].Title)
+}
+
+func TestCreateDatabasePackage(t *testing.T) {
+	t.Parallel()
+
+	cases := []struct {
+		title    string
+		cursor   string
+		pkgBytes []byte
+		expected *database.Package
+	}{
+		{
+			title: "package with minimum data",
+			pkgBytes: []byte(`{
+  "name": "mypackage",
+  "version": "1.2.3",
+  "format_version": "2.2.2",
+  "type": "integration",
+  "description": "My package description",
+  "categories": ["cat1", "cat2"],
+  "conditions": {
+    "kibana": {
+      "version": "^8.17.0"
+	}
+  }
+}`),
+			cursor: "1",
+			expected: &database.Package{
+				Cursor:                  "1",
+				Name:                    "mypackage",
+				Version:                 "1.2.3",
+				KibanaVersion:           "^8.17.0",
+				FormatVersion:           "2.2.2",
+				FormatVersionMajorMinor: "2.2.0",
+				Type:                    "integration",
+				Path:                    "mypackage-1.2.3.zip",
+				Data:                    []byte(`{"name":"mypackage","version":"1.2.3","description":"My package description","type":"integration","download":"","path":"","conditions":{"kibana":{"version":"^8.17.0"}},"categories":["cat1","cat2"],"format_version":"2.2.2"}`),
+				BaseData:                []byte(`{"name":"mypackage","version":"1.2.3","description":"My package description","type":"integration","download":"","path":"","conditions":{"kibana":{"version":"^8.17.0"}},"categories":["cat1","cat2"]}`),
+				Prerelease:              false,
+			},
+		},
+		{
+			title: "prerelease package",
+			pkgBytes: []byte(`{
+  "name": "mypackage",
+  "version": "1.2.3-beta1",
+  "format_version": "2.2.2",
+  "type": "integration",
+  "description": "My package description",
+  "categories": ["cat1", "cat2"],
+  "conditions": {
+    "kibana": {
+      "version": "^8.17.0"
+	}
+  }
+}`),
+			cursor: "1",
+			expected: &database.Package{
+				Cursor:                  "1",
+				Name:                    "mypackage",
+				Version:                 "1.2.3-beta1",
+				KibanaVersion:           "^8.17.0",
+				FormatVersion:           "2.2.2",
+				FormatVersionMajorMinor: "2.2.0",
+				Type:                    "integration",
+				Path:                    "mypackage-1.2.3-beta1.zip",
+				Data:                    []byte(`{"name":"mypackage","version":"1.2.3-beta1","description":"My package description","type":"integration","download":"","path":"","conditions":{"kibana":{"version":"^8.17.0"}},"categories":["cat1","cat2"],"format_version":"2.2.2"}`),
+				BaseData:                []byte(`{"name":"mypackage","version":"1.2.3-beta1","description":"My package description","type":"integration","download":"","path":"","conditions":{"kibana":{"version":"^8.17.0"}},"categories":["cat1","cat2"]}`),
+				Prerelease:              true,
+			},
+		},
+	}
+	for _, c := range cases {
+		t.Run(c.title, func(t *testing.T) {
+			// when
+			pkg := &packages.Package{}
+			err := json.Unmarshal(c.pkgBytes, pkg)
+			require.NoError(t, err, "package should be unmarshalled")
+			dbPkg, err := createDatabasePackage(pkg, c.cursor)
+			require.NoError(t, err, "database package should be created")
+			// then
+			assert.Equal(t, c.expected, dbPkg)
+		})
+	}
 }