This PR adds new functions into SQL to be able to compare fields using semantic version:
- semver_compare_constraint
- semver_compare_ge
- semver_compare_le

These functions allow the indexer to retrieve less packages when query parameters like `kibana.version`, `spec.min` or `spec.max` are used.

- Benchmark:
```
 $ benchstat sql_prod_indexer_main.txt sql_prod_indexer_functions.txt 
goos: linux
goarch: amd64
pkg: github.com/elastic/package-registry/internal/storage
cpu: 11th Gen Intel(R) Core(TM) i7-11800H @ 2.30GHz
                 │ sql_prod_indexer_main.txt │   sql_prod_indexer_functions.txt   │
                 │          sec/op           │   sec/op    vs base                │
SQLIndexerGet-16                  2.018 ± 2%   1.607 ± 2%  -20.35% (p=0.000 n=10)

                 │ sql_prod_indexer_main.txt │    sql_prod_indexer_functions.txt    │
                 │           B/op            │     B/op      vs base                │
SQLIndexerGet-16                520.1Mi ± 0%   414.1Mi ± 0%  -20.37% (p=0.000 n=10)

                 │ sql_prod_indexer_main.txt │   sql_prod_indexer_functions.txt    │
                 │         allocs/op         │  allocs/op   vs base                │
SQLIndexerGet-16                 7.281M ± 0%   5.863M ± 0%  -19.48% (p=0.000 n=10)
```
- Comparison

| Query | Main (no functions) | Using new functions | 
| --- | --- | --- | 
| `/search?kibana.version=9.0.0` | 8898 | 1345 |
| `/search?spec.min=2.0&spec.max=3.4` | 8898 | 5733 |

## Author's checklist 

- [x] Run benchmark to analyze the benefits.
- [x] Compare number of packages retrieved.

## How to test this locally

Launch Package Registry service:
```shell
bash dev/launch_epr_service_storage_indexer.sh -s -i storage/testdata/search-index-all-full.json -p localhost:8081 -d
```

Run some queries against that service launched previously, some examples:
```shell
curl -s "http://localhost:8081/search?kibana.version=8.16.0"
curl -s "http://localhost:8081/search?kibana.version=7.17.0"

curl -s "http://localhost:8081/search?spec.min=2.10&spec.max=3.1"
```