@mrodm 2025-10-07
> > Latest benchmark results obtained with the changes introduced in this PR:
> 
> Are these benchmarks compared with the previous version or with main? They look worse than the ones in the description, right?

> Are these benchmarks compared with the previous version or with main? They look worse than the ones in the description, right?

Those were compared with the contents in the main branch. It's right that they are a bit worse. I cannot tell you exactly why it could be, but these benchmarks as they are run in my machine, maybe some other program was interfering. At least, in general there are some benefits.

This is a new benchmark repeated locally in my machine for both main and this branch:
```
goos: linux
goarch: amd64
pkg: github.com/elastic/package-registry/internal/storage
cpu: 11th Gen Intel(R) Core(TM) i7-11800H @ 2.30GHz
                 │ ../package-registry-sql-queries/sql_prod_get_main_3.txt │ sql_prod_get_compare_functions_3.txt │
                 │                         sec/op                          │    sec/op      vs base               │
SQLIndexerGet-16                                               2.313 ± 27%      2.148 ± 1%  -7.14% (p=0.023 n=10)

                 │ ../package-registry-sql-queries/sql_prod_get_main_3.txt │ sql_prod_get_compare_functions_3.txt │
                 │                          B/op                           │     B/op       vs base               │
SQLIndexerGet-16                                              437.8Mi ± 0%    413.0Mi ± 0%  -5.67% (p=0.000 n=10)

                 │ ../package-registry-sql-queries/sql_prod_get_main_3.txt │ sql_prod_get_compare_functions_3.txt │
                 │                        allocs/op                        │   allocs/op    vs base               │
SQLIndexerGet-16                                               5.873M ± 0%     5.694M ± 0%  -3.05% (p=0.000 n=10)
```